refactor: consolidate pending result API routes
