+++ src/app/components/SubmissionPopUp.tsx
@@ -24,7 +24,7 @@ const SubmissionPopUp = ({ isOpen, onClose, result }: PopUpProps) => {
 
   async function submitResult(result: Result) {
     try {
-      const response = await fetch(`/api/pending/post`, {
+      const response = await fetch(`/api/pending`, {
         method: "POST",
         headers: {
           "Content-Type": "application/json",
